Improve AttributeData display
